Crate common_catalog

Modules

consts

Functions

build_db_string

Build db name from catalog and schema string

format_full_flow_name

Formats flow fully-qualified name

format_full_table_name

Formats table fully-qualified name

format_schema_name

parse_catalog_and_schema_from_db_string

Attempt to parse catalog and schema from given database name

parse_optional_catalog_and_schema_from_db_string

Attempt to parse catalog and schema from given database name